+++ src/private.rs
@@ -9,9 +9,9 @@
 
 use crate::chan::{Ch16, Ch32, Ch64, Ch8, Channel};
 use crate::mono::Mono;
-use crate::sample::{Sample1, Sample2, Sample6, Sample8};
+use crate::sample::{Sample1, Sample2, Sample4, Sample6, Sample8};
 use crate::stereo::Stereo;
-use crate::surround::{Surround, SurroundHD};
+use crate::surround::{Surround4, Surround5, Surround7};
 use crate::Config;
 use core::any::Any;
 
@@ -22,9 +22,11 @@ impl Sealed for Ch32 {}
 impl Sealed for Ch64 {}
 impl Sealed for Mono {}
 impl Sealed for Stereo {}
-impl Sealed for Surround {}
-impl Sealed for SurroundHD {}
+impl Sealed for Surround4 {}
+impl Sealed for Surround5 {}
+impl Sealed for Surround7 {}
 impl<C: Channel, F: Config> Sealed for Sample1<C, F> {}
 impl<C: Channel, F: Config> Sealed for Sample2<C, F> {}
+impl<C: Channel, F: Config> Sealed for Sample4<C, F> {}
 impl<C: Channel, F: Config> Sealed for Sample6<C, F> {}
 impl<C: Channel, F: Config> Sealed for Sample8<C, F> {}